+++ cloud_function/util/simulation.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from typing import Optional
 
 
 class Simulation:
